+++ internal/http/http_test.go
@@ -1,3 +1,8 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the MIT License.
+// This product includes software developed at Guance Cloud (https://www.guance.com/).
+// Copyright 2021-present Guance, Inc.
+
 package http
 
 import (
@@ -430,12 +435,15 @@ type slowReader struct {
 	buf *bytes.Buffer
 }
 
+var sleepms = time.Duration(1000)
+
 func (r *slowReader) Read(p []byte) (int, error) {
 	if r.buf == nil {
 		return 0, nil
 	}
 
-	time.Sleep(time.Millisecond * 100) // slow reader
+	// slow reader
+	time.Sleep(sleepms * time.Millisecond) //nolint:durationcheck
 	return r.buf.Read(p)
 }
 
@@ -448,8 +456,8 @@ func TestServerTimeout(t *testing.T) {
 	})
 
 	ts := httptest.NewUnstartedServer(r)
-	ts.Config.ReadTimeout = 1000 * time.Millisecond // easy to timeout
-	ts.Config.WriteTimeout = 20 * time.Millisecond  // easy to timeout
+	// ts.Config.ReadTimeout = sleepms * time.Millisecond // nolint:durationcheck
+	ts.Config.WriteTimeout = (sleepms * 10) * time.Millisecond //nolint:durationcheck
 	ts.Start()
 
 	defer ts.Close()
@@ -459,16 +467,17 @@ func TestServerTimeout(t *testing.T) {
 		// Timeout: time.Second, not set
 	}
 
-	req, err := http.NewRequest("GET", ts.URL+"/test",
-		&slowReader{buf: bytes.NewBufferString("body string")})
+	// req, err := http.NewRequest("GET", ts.URL+"/test", &slowReader{buf: bytes.NewBufferString("body string")})
+	req, err := http.NewRequest("GET", ts.URL+"/test", nil)
 	if err != nil {
 		t.Error(err)
 		return
 	}
 
+	start := time.Now()
 	resp, err := cli.Do(req)
 	if err != nil {
-		t.Logf("Do: %s, type: %s, %+#v, Err: %+#v", err, reflect.TypeOf(err), err, err.(*url.Error).Err) //nolint:errorlint
+		t.Logf("Do: %s\ntype: %s, %+#v, Err: %+#v\ncost: %s", err, reflect.TypeOf(err), err, err.(*url.Error).Err, time.Since(start)) //nolint:errorlint
 	} else {
 		defer resp.Body.Close()
 	}