+++ io/dataway/dataway_default_test.go
@@ -1,3 +1,8 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the MIT License.
+// This product includes software developed at Guance Cloud (https://www.guance.com/).
+// Copyright 2021-present Guance, Inc.
+
 package dataway
 
 import (
@@ -16,9 +21,10 @@ import (
 )
 
 func TestDataWayAPIs(t *testing.T) {
-	dw := DataWayCfg{URLs: []string{"https://abc.com?token=tkn_abc"}}
+	dwCfg := DataWayCfg{URLs: []string{"https://abc.com?token=tkn_abc"}}
 
-	if err := dw.Apply(); err != nil {
+	dw := &DataWayDefault{}
+	if err := dw.Init(&dwCfg); err != nil {
 		t.Fatal(err)
 	}
 
@@ -45,8 +51,9 @@ func TestHeartBeat(t *testing.T) {
 	}
 
 	for _, tc := range cases {
-		dw := &DataWayCfg{URLs: tc.urls, ontest: true}
-		err := dw.Apply()
+		dwCfg := &DataWayCfg{URLs: tc.urls}
+		dw := &DataWayDefault{ontest: true}
+		err := dw.Init(dwCfg)
 		tu.Equals(t, nil, err)
 
 		_, err = dw.HeartBeat()
@@ -78,8 +85,9 @@ func TestListDataWay(t *testing.T) {
 	}
 
 	for _, tc := range cases {
-		dw := &DataWayCfg{URLs: tc.urls, ontest: true}
-		err := dw.Apply()
+		dwCfg := &DataWayCfg{URLs: tc.urls}
+		dw := &DataWayDefault{ontest: true}
+		err := dw.Init(dwCfg)
 		tu.Equals(t, nil, err)
 
 		dws, _, err := dw.DatawayList()
@@ -125,13 +133,14 @@ func TestSend(t *testing.T) {
 	for idx, tc := range cases {
 		t.Logf("===== case %d ======", idx)
 
-		dw := &DataWayCfg{URLs: tc.urls}
-		if err := dw.Apply(); err != nil {
+		dwCfg := &DataWayCfg{URLs: tc.urls}
+		dw := &DataWayDefault{ontest: true}
+		if err := dw.Init(dwCfg); err != nil {
 			t.Errorf("Apply: %s", err.Error())
 			continue
 		}
 
-		err := dw.Send(tc.category, []byte("abc123"), tc.gz)
+		_, err := dw.Send(tc.category, []byte("abc123"), tc.gz)
 		if tc.fail {
 			tu.NotOk(t, err, "")
 		} else {
@@ -159,8 +168,9 @@ func TestElectionHeartBeatURL(t *testing.T) {
 	}
 
 	for _, tc := range cases {
-		dw := DataWayCfg{URLs: tc.urls}
-		err := dw.Apply()
+		dwCfg := &DataWayCfg{URLs: tc.urls}
+		dw := &DataWayDefault{}
+		err := dw.Init(dwCfg)
 		if tc.fail {
 			tu.NotOk(t, err, "")
 		} else {
@@ -197,8 +207,9 @@ func TestElectionURL(t *testing.T) {
 	}
 
 	for _, tc := range cases {
-		dw := DataWayCfg{URLs: tc.urls}
-		err := dw.Apply()
+		dwCfg := &DataWayCfg{URLs: tc.urls}
+		dw := &DataWayDefault{}
+		err := dw.Init(dwCfg)
 		if tc.fail {
 			tu.NotOk(t, err, "")
 		} else {
@@ -243,8 +254,9 @@ func TestGetToken(t *testing.T) {
 	}
 
 	for _, tc := range cases {
-		dw := DataWayCfg{URLs: tc.urls}
-		err := dw.Apply()
+		dwCfg := &DataWayCfg{URLs: tc.urls}
+		dw := &DataWayDefault{}
+		err := dw.Init(dwCfg)
 		if tc.fail {
 			tu.NotOk(t, err, "")
 			continue
@@ -252,7 +264,7 @@ func TestGetToken(t *testing.T) {
 			tu.Ok(t, err)
 		}
 
-		tkns := dw.GetToken()
+		tkns := dw.GetTokens()
 		for idx, x := range tkns {
 			tu.Equals(t, tc.expect[idx], x)
 		}
@@ -311,14 +323,16 @@ func TestSetupDataway(t *testing.T) {
 	for i, tc := range cases {
 		t.Logf("case %d...", i)
 
-		dw := DataWayCfg{
+		dwCfg := &DataWayCfg{
 			DeprecatedURL: tc.url,
 			URLs:          tc.urls,
 			HTTPProxy:     tc.proxy,
 			Proxy:         tc.proxy != "",
 		}
 
-		err := dw.Apply()
+		dw := &DataWayDefault{}
+		err := dw.Init(dwCfg)
+
 		if tc.fail {
 			tu.NotOk(t, err, "")
 			continue
@@ -369,15 +383,16 @@ func runTestDatawayConnections(t *testing.T, nreq int) {
 	ts.Start()
 	defer ts.Close()
 
-	dw := &DataWayCfg{URLs: []string{ts.URL}}
-	if err := dw.Apply(); err != nil {
+	dwCfg := &DataWayCfg{URLs: []string{ts.URL}}
+	dw := &DataWayDefault{}
+	if err := dw.Init(dwCfg); err != nil {
 		t.Fatal(err)
 	}
 
 	t.Logf("dw: %+#v", dw)
 
 	for {
-		if err := dw.Send("/v1/write/metric", []byte("abc123"), false); err != nil {
+		if _, err := dw.Send("/v1/write/metric", []byte("abc123"), false); err != nil {
 			t.Fatal(err)
 		}
 
@@ -397,8 +412,9 @@ func TestUploadLog(t *testing.T) {
 	}))
 	defer ts.Close()
 
-	dw := &DataWayCfg{URLs: []string{ts.URL}}
-	if err := dw.Apply(); err != nil {
+	dwCfg := &DataWayCfg{URLs: []string{ts.URL}}
+	dw := &DataWayDefault{}
+	if err := dw.Init(dwCfg); err != nil {
 		t.Errorf("Apply: %s", err.Error())
 	}
 	rBody := strings.NewReader("aaaaaaaaaaaaa")
@@ -412,17 +428,22 @@ func TestUploadLog(t *testing.T) {
 
 func TestDatawayTimeout(t *testing.T) {
 	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		time.Sleep(time.Second * 3) // timeout
+		time.Sleep(time.Second * 1) // timeout
 		fmt.Fprint(w, "OK")
 	}))
 	defer ts.Close()
 
-	dw := &DataWayCfg{URLs: []string{ts.URL}, HTTPTimeout: "1s"}
+	dw := &DataWayDefault{
+		DataWayCfg: &DataWayCfg{
+			URLs:        []string{ts.URL},
+			HTTPTimeout: "100ms",
+		},
+	}
 	if err := dw.Apply(); err != nil {
 		t.Errorf("Apply: %s", err.Error())
 	}
 
-	t.Logf("http client timeout: %s", dw.httpCli.Timeout)
+	t.Logf("http client timeout: %s", dw.httpCli.HTTPClient.Timeout)
 
 	ch := make(chan interface{})
 	go func() {
@@ -433,7 +454,9 @@ func TestDatawayTimeout(t *testing.T) {
 		close(ch)
 	}()
 
-	tick := time.NewTicker(time.Second * 2)
+	// for timeout 3 times,1st time, wait for 1s(server sleep 1s) + 1s(1st retry) + 2s(2nd retry)
+	// 1st send + 1st retry + 2nd retry = 3 times
+	tick := time.NewTicker(time.Second * 5)
 	defer tick.Stop()
 	select {
 	case <-ch:
@@ -455,7 +478,7 @@ func TestCheckToken(t *testing.T) {
 		{valid: false, token: "token_xxxxxxxxx"},
 		{valid: false, token: "tkn_xxxxxxxxx"},
 	}
-	dw := DataWayCfg{}
+	dw := DataWayDefault{}
 
 	for _, info := range cases {
 		err := dw.CheckToken(info.token)