+++ io/dataway/send.go
@@ -1,22 +1,35 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the MIT License.
+// This product includes software developed at Guance Cloud (https://www.guance.com/).
+// Copyright 2021-present Guance, Inc.
+
 package dataway
 
 import (
 	"bytes"
 	"context"
+	"crypto/tls"
+	"errors"
 	"fmt"
 	"io/ioutil"
 	"math/rand"
 	"net/http"
+	"net/http/httptrace"
 	"net/url"
 	"sync"
 	"time"
 
+	"github.com/hashicorp/go-retryablehttp"
+	"gitlab.jiagouyun.com/cloudcare-tools/datakit"
+	"gitlab.jiagouyun.com/cloudcare-tools/datakit/io/sender"
+	"gitlab.jiagouyun.com/cloudcare-tools/datakit/io/sink/sinkcommon"
 	"gopkg.in/CodapeWild/dd-trace-go.v1/ddtrace/ext"
 	"gopkg.in/CodapeWild/dd-trace-go.v1/ddtrace/tracer"
 )
 
 var (
 	sendFailStats = map[string]int32{}
+	startTime     = time.Now()
 	lock          sync.RWMutex
 )
 
@@ -50,10 +63,11 @@ func updateSendFailStats(category string, isOk bool) {
 	log.Debugf("update send fail stats: %+#v", sendFailStats)
 }
 
-func (dc *endPoint) send(category string, data []byte, gz bool) error {
+func (dc *endPoint) send(category string, data []byte, gz bool) (int, error) {
 	var (
-		err      error
-		isSendOk bool // data sent successfully, http response code is 200
+		err        error
+		isSendOk   bool // data sent successfully, http response code is 200
+		statusCode int
 	)
 
 	span, _ := tracer.StartSpanFromContext(context.Background(), "io.dataway.send", tracer.SpanType(ext.SpanTypeHTTP))
@@ -74,7 +88,7 @@ func (dc *endPoint) send(category string, data []byte, gz bool) error {
 
 		// for dialtesting, there are user-defined url to post
 		if x, err := url.ParseRequestURI(category); err != nil {
-			return fmt.Errorf("invalid url %s", category)
+			return statusCode, fmt.Errorf("invalid url %s", category)
 		} else {
 			log.Debugf("try use URL %+#v", x)
 			requrl = category
@@ -85,7 +99,7 @@ func (dc *endPoint) send(category string, data []byte, gz bool) error {
 	if req, err = http.NewRequest("POST", requrl, bytes.NewBuffer(data)); err != nil {
 		log.Error(err)
 
-		return err
+		return statusCode, err
 	}
 	span.SetTag("method", req.Method)
 	span.SetTag("url", requrl)
@@ -100,7 +114,7 @@ func (dc *endPoint) send(category string, data []byte, gz bool) error {
 	if dc.ontest {
 		log.Debug("Datakit client on test")
 
-		return nil
+		return statusCode, nil
 	}
 
 	tracer.Inject(span.Context(), tracer.HTTPHeadersCarrier(req.Header)) //nolint:errcheck,gosec
@@ -113,7 +127,13 @@ func (dc *endPoint) send(category string, data []byte, gz bool) error {
 		dc.fails++
 		log.Errorf("request url %s failed(proxy: %s): %s", requrl, dc.proxy, err)
 
-		return err
+		var urlError *url.Error
+
+		if errors.As(err, &urlError) && urlError.Timeout() {
+			statusCode = -1 // timeout
+		}
+
+		return statusCode, err
 	}
 	span.SetTag("status", resp.Status)
 
@@ -122,7 +142,7 @@ func (dc *endPoint) send(category string, data []byte, gz bool) error {
 	if body, err = ioutil.ReadAll(resp.Body); err != nil {
 		log.Error(err)
 
-		return err
+		return statusCode, err
 	}
 
 	switch resp.StatusCode / 100 {
@@ -150,18 +170,93 @@ func (dc *endPoint) send(category string, data []byte, gz bool) error {
 		err = fmt.Errorf("dataway internal error")
 	}
 
-	return err
+	return statusCode, err
 }
 
-func (dw *DataWayCfg) sendReq(req *http.Request) (*http.Response, error) {
+type httpTraceStat struct {
+	reuseConn bool
+	idle      bool
+	idleTime  time.Duration
+
+	dnsStart   time.Time
+	dnsResolve time.Duration
+	tlsHSStart time.Time
+	tlsHSDone  time.Duration
+	connStart  time.Time
+	connDone   time.Duration
+	ttfbTime   time.Duration
+
+	cost time.Duration
+}
+
+func (ts *httpTraceStat) String() string {
+	if ts == nil {
+		return "-"
+	}
+
+	return fmt.Sprintf("dataway httptrace: Conn: [reuse: %v,idle: %v/%s], DNS: %s, TLS: %s, Connect: %s, TTFB: %s, cost: %s",
+		ts.reuseConn, ts.idle, ts.idleTime, ts.dnsResolve, ts.tlsHSDone, ts.connDone, ts.ttfbTime, ts.cost)
+}
+
+type DatawayError struct {
+	Err   error
+	Trace *httpTraceStat
+	API   string
+}
+
+func (de *DatawayError) Error() string {
+	return fmt.Sprintf("HTTP error: %s, API: %s, httptrace: %s",
+		de.Err, de.API, de.Trace)
+}
+
+func (dw *DataWayDefault) sendReq(req *http.Request) (*http.Response, error) {
 	log.Debugf("send request %s, proxy: %s, dwcli: %p, timeout: %s(%s)",
-		req.URL.String(), dw.HTTPProxy, dw.httpCli.Transport,
+		req.URL.String(), dw.HTTPProxy, dw.httpCli.HTTPClient.Transport,
 		dw.HTTPTimeout, dw.TimeoutDuration.String())
 
-	return dw.httpCli.Do(req)
+	var reqStart time.Time
+	var ts *httpTraceStat
+	if dw.EnableHTTPTrace {
+		ts = &httpTraceStat{}
+		t := &httptrace.ClientTrace{
+			GotConn: func(ci httptrace.GotConnInfo) {
+				ts.reuseConn = ci.Reused
+				ts.idle = ci.WasIdle
+				ts.idleTime = ci.IdleTime
+			},
+			DNSStart:             func(httptrace.DNSStartInfo) { ts.dnsStart = time.Now() },
+			DNSDone:              func(httptrace.DNSDoneInfo) { ts.dnsResolve = time.Since(ts.dnsStart) },
+			TLSHandshakeStart:    func() { ts.tlsHSStart = time.Now() },
+			TLSHandshakeDone:     func(tls.ConnectionState, error) { ts.tlsHSDone = time.Since(ts.tlsHSStart) },
+			ConnectStart:         func(string, string) { ts.connStart = time.Now() },
+			ConnectDone:          func(string, string, error) { ts.connDone = time.Since(ts.connStart) },
+			GotFirstResponseByte: func() { ts.ttfbTime = time.Since(reqStart) },
+		}
+
+		req = req.WithContext(httptrace.WithClientTrace(req.Context(), t))
+	}
+
+	reqStart = time.Now()
+	x, err := retryablehttp.FromRequest(req)
+	if err != nil {
+		log.Errorf("retryablehttp.FromRequest: %s", err)
+		return nil, err
+	}
+
+	resp, err := dw.httpCli.Do(x)
+	if ts != nil {
+		ts.cost = time.Since(reqStart)
+		log.Debugf("%s: %s", req.URL.Path, ts.String())
+	}
+
+	if err != nil {
+		return nil, &DatawayError{Err: err, Trace: ts, API: req.URL.Path}
+	}
+
+	return resp, nil
 }
 
-func (dw *DataWayCfg) Send(category string, data []byte, gz bool) error {
+func (dw *DataWayDefault) Send(category string, data []byte, gz bool) (statusCode int, err error) {
 	for i, ep := range dw.endPoints {
 		log.Debugf("send to %dth dataway, fails: %d/%d", i, ep.fails, dw.MaxFails)
 		// 判断 fails
@@ -169,21 +264,122 @@ func (dw *DataWayCfg) Send(category string, data []byte, gz bool) error {
 			rand.Seed(time.Now().UnixNano())
 			index := rand.Intn(len(AvailableDataways)) //nolint:gosec
 
-			var err error
 			url := fmt.Sprintf(`%s?%s`, AvailableDataways[index], ep.urlValues.Encode())
 			ep, err = dw.initEndpoint(url)
 			if err != nil {
 				log.Error(err)
-				return err
+				return
 			}
 
 			dw.endPoints[i] = ep
 		}
 
-		if err := ep.send(category, data, gz); err != nil {
-			return err
+		statusCode, err = ep.send(category, data, gz)
+		if err != nil {
+			return
 		}
 	}
 
+	return
+}
+
+func (dw *DataWayDefault) Write(category string, pts []sinkcommon.ISinkPoint) error {
+	if len(pts) == 0 {
+		return nil
+	}
+
+	var statusCode int
+
+	bodies, err := dw.buildBody(pts, true)
+	if err != nil {
+		return err
+	}
+
+	rawBytes := 0
+	gzBytes := 0
+
+	for _, body := range bodies {
+		if code, err := dw.Send(category, body.buf, body.gzon); err != nil {
+			log.Error(err)
+			sender.FeedMetric(&sender.SinkMetric{
+				Name:       "dataway",
+				StartTime:  startTime,
+				IsSuccess:  false,
+				StatusCode: code,
+			})
+			return err
+		} else {
+			rawBytes += int(body.rawBufBytes)
+			gzBytes += len(body.buf)
+			statusCode = code
+		}
+	}
+
+	sender.FeedMetric(&sender.SinkMetric{
+		Name:       "dataway",
+		IsSuccess:  true,
+		StartTime:  startTime,
+		Pts:        uint64(len(pts)),
+		Bytes:      uint64(gzBytes),
+		RawBytes:   uint64(rawBytes),
+		StatusCode: statusCode,
+	})
+
 	return nil
+}
+
+const (
+	minGZSize   = 1024
+	maxKodoPack = 10 * 1000 * 1000
+)
+
+type body struct {
+	buf         []byte
+	gzon        bool
+	rawBufBytes int64
+}
+
+func (dw *DataWayDefault) buildBody(pts []sinkcommon.ISinkPoint, isGzip bool) ([]*body, error) {
+	lines := bytes.Buffer{}
+	var (
+		gz = func(lines []byte) (*body, error) {
+			var (
+				body = &body{buf: lines, rawBufBytes: int64(len(lines))}
+				err  error
+			)
+			log.Debugf("### io body size before GZ: %dM %dK", len(body.buf)/1000/1000, len(body.buf)/1000)
+			if len(lines) > minGZSize && isGzip {
+				if body.buf, err = datakit.GZip(body.buf); err != nil {
+					log.Errorf("gz: %s", err.Error())
+
+					return nil, err
+				}
+				body.gzon = true
+			}
+
+			return body, nil
+		}
+		// lines  bytes.Buffer
+		bodies []*body
+	)
+	lines.Reset()
+	for _, pt := range pts {
+		ptstr := pt.String()
+		if lines.Len()+len(ptstr)+1 >= maxKodoPack {
+			if body, err := gz(lines.Bytes()); err != nil {
+				return nil, err
+			} else {
+				log.Warn(string(body.buf))
+				bodies = append(bodies, body)
+			}
+			lines.Reset()
+		}
+		lines.WriteString(ptstr)
+		lines.WriteString("\n")
+	}
+	if body, err := gz(lines.Bytes()); err != nil {
+		return nil, err
+	} else {
+		return append(bodies, body), nil
+	}
 }