+++ plugins/inputs/container/input.go
@@ -1,3 +1,8 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the MIT License.
+// This product includes software developed at Guance Cloud (https://www.guance.com/).
+// Copyright 2021-present Guance, Inc.
+
 // Package container collect container metrics/loggings/objects.
 package container
 
@@ -194,6 +199,10 @@ func (i *Input) collectMetric() {
 		l.Errorf("failed to watch container log: %w", err)
 	}
 
+	if err := i.watchNewContainerdLogs(); err != nil {
+		l.Errorf("failed to watch containerd log: %w", err)
+	}
+
 	if !datakit.Docker {
 		return
 	}
@@ -293,6 +302,13 @@ func (i *Input) gatherContainerdObject() error {
 
 	return inputs.FeedMeasurement("containerd-object", datakit.Object, res,
 		&io.Option{CollectCost: time.Since(start)})
+}
+
+func (i *Input) watchNewContainerdLogs() error {
+	if i.containerdInput == nil {
+		return nil
+	}
+	return i.containerdInput.watchNewLogs()
 }
 
 func (i *Input) gatherK8sResourceMetric() error {