+++ plugins/inputs/container/k8s_pod.go
@@ -1,8 +1,14 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the MIT License.
+// This product includes software developed at Guance Cloud (https://www.guance.com/).
+// Copyright 2021-present Guance, Inc.
+
 package container
 
 import (
 	"context"
 	"fmt"
+	"os"
 	"time"
 
 	"gitlab.jiagouyun.com/cloudcare-tools/datakit"
@@ -128,7 +134,6 @@ func (p *pod) object() (inputsMeas, error) {
 				"name":         fmt.Sprintf("%v", item.UID),
 				"pod_name":     item.Name,
 				"node_name":    item.Spec.NodeName,
-				"host":         item.Spec.NodeName, // 指定 pod 所在的 host
 				"phase":        fmt.Sprintf("%v", item.Status.Phase),
 				"qos_class":    fmt.Sprintf("%v", item.Status.QOSClass),
 				"state":        fmt.Sprintf("%v", item.Status.Phase), // Depercated
@@ -142,6 +147,10 @@ func (p *pod) object() (inputsMeas, error) {
 				"create_time": item.CreationTimestamp.Time.Unix(),
 			},
 			time: time.Now(),
+		}
+
+		if n := getHostname(); n != "" {
+			obj.tags["host"] = n // 指定 pod 所在的 host
 		}
 
 		for _, ref := range item.OwnerReferences {
@@ -331,6 +340,14 @@ func (*podObject) Info() *inputs.MeasurementInfo {
 			"message":     &inputs.FieldInfo{DataType: inputs.String, Unit: inputs.UnknownUnit, Desc: "object details"},
 		},
 	}
+}
+
+func getHostname() string {
+	if e := os.Getenv("ENV_K8S_NODE_NAME"); e != "" {
+		return e
+	}
+	n, _ := os.Hostname()
+	return n
 }
 
 //nolint:gochecknoinits